restructured component layout
